Fixed and optimized word set generator

+++ tasks/WordCount/WordMapper.java
@@ -2,14 +2,13 @@
 import java.io.IOException;
 import java.util.StringTokenizer;
 
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
 public class WordMapper {
-	public static class Map extends Mapper<LongWritable, Text, Text, IntWritable> {
-		private final static IntWritable one = new IntWritable(1);
+	public static class Map extends Mapper<LongWritable, Text, Text, LongWritable> {
+		private final static LongWritable one = new LongWritable(1);
 		private Text word = new Text();
 
 		public void map(LongWritable key, Text value, Context context)
